requirements.txt
# 短线稳健策略系统依赖
akshare>=1.12.0
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0
streamlit>=1.28.0  # 可选，用于Web界面
plotly>=5.18.0     # 可选，用于可视化
